requirements.txt
fastapi
uvicorn
requests
pydantic
tenacity
beautifulsoup4
python-dotenv
